requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pandas-ta>=0.3.14b0
finnhub-python>=2.4.13
requests>=2.26.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
matplotlib>=3.4.3
pytz>=2021.1
scipy>=1.9.0
joblib>=1.0.0
nltk>=3.6.0
textblob>=0.15.3
pytest>=7.0.0
pytest-cov>=3.0.0
tqdm>=4.62.0
openpyxl>=3.0.9
seaborn>=0.11.2
statsmodels>=0.13.0
websocket-client>=1.4.0 